Move span kind assignment to OpenTelemetry patch
