+++ test/new_relic/agent/span_event_primitive_test.rb
@@ -408,43 +408,6 @@ module NewRelic
           assert_equal thread_id, spans[1][0][0]['thread.id']
         end
 
-        def test_span_kind_added_to_intrinsics_when_otel_span_has_kind
-          in_transaction do |txn|
-            txn.stubs(:sampled?).returns(true)
-            segment = txn.current_segment
-            otel_span = stub(kind: :server)
-            segment.instance_variable_set(:@otel_span, otel_span)
-
-            intrinsics, = SpanEventPrimitive.for_segment(segment)
-
-            assert_equal 'server', intrinsics['span.kind']
-          end
-        end
-
-        def test_span_kind_not_added_to_intrinsics_without_otel_span
-          in_transaction do |txn|
-            txn.stubs(:sampled?).returns(true)
-            segment = txn.current_segment
-
-            intrinsics, = SpanEventPrimitive.for_segment(segment)
-
-            refute intrinsics.key?('span.kind')
-          end
-        end
-
-        def test_span_kind_not_added_to_intrinsics_when_otel_span_kind_is_nil
-          in_transaction do |txn|
-            txn.stubs(:sampled?).returns(true)
-            segment = txn.current_segment
-            otel_span = stub(kind: nil)
-            segment.instance_variable_set(:@otel_span, otel_span)
-
-            intrinsics, = SpanEventPrimitive.for_segment(segment)
-
-            refute intrinsics.key?('span.kind')
-          end
-        end
-
         def test_for_segment_returns_three_element_array_when_no_links
           with_segment do |segment|
             event = SpanEventPrimitive.for_segment(segment)
